// optimizer/types/all-positions.interface.ts
import { Position } from "./position.interface";

export interface AllPositions {
    ST: Position;
    LF: Position;
    RF: Position;
    CF: Position;
    LW: Position;
    RW: Position;
    LM: Position;
    RM: Position;
    CAM: Position;
    CM: Position;
    CDM: Position;
    LWB: Position;
    RWB: Position;
    LB: Position;
    RB: Position;
    CB: Position;
    GK: Position;
}